ios and ubuntu

+++ examples/phpsciter-load-frame/index.phpt
@@ -17,14 +17,19 @@ if ($pid == 0)
     PHPSciter::ALLOW_FILE_IO | PHPSciter::ALLOW_SOCKET_IO | PHPSciter::ALLOW_EVAL |
                                    PHPSciter::ALLOW_SYSINFO);
 
-
+    exit(3);
     $oSciter->run(PHPSciter::SW_TITLEBAR | PHPSciter::SW_RESIZEABLE | PHPSciter::SW_MAIN | PHPSciter::SW_ENABLE_DEBUG
     |PHPSciter::SW_CONTROLS);
     exit(0);
 } else {
-    sleep(1);
+    sleep(2);
+    pcntl_waitpid($pid, $status, WNOHANG);
+    if (pcntl_wexitstatus($status) > 0)
+    {
+        exit(pcntl_wexitstatus($status));
+    }
     $r = posix_kill($pid, 15);
-    var_dump($pid);
+
 }
 
 ?>
